Finishing up a minor refactor

+++ dpu/utils.py
@@ -95,32 +95,42 @@ def rsync(source, target, excludes=None):
     subprocess.check_call(cmd, shell=False)
 
 
-def run_builder(binary, path):
-    binary = abspath("./builders/%s" % (binary))
-    path = abspath(path)
+def run_builder(biny, path):
+    binaries = [
+        abspath("./builders/%s" % (biny)),
+        abspath("/usr/lib/dpu/builders/%s" % (biny))
+    ]
+    for binary in binaries:
+        if os.path.exists(binary):
+            null = open("/dev/null", "w")
+            stdout = null
+            stderr = null
 
-    null = open("/dev/null", "w")
-    stdout = null
-    stderr = null
-
-    cmd = [binary, path]
-    subprocess.check_call(cmd,
-                          shell=False,
-                          stdout=stdout,
-                          stderr=stderr)
+            cmd = [binary, path]
+            subprocess.check_call(cmd,
+                                  shell=False,
+                                  stdout=stdout,
+                                  stderr=stderr)
+            return
+    raise Exception("No such builder: %s" % (biny))
 
 
-def run_checker(binary, path):
-    binary = abspath("./checkers/%s" % (binary))
-    path = abspath(path)
+def run_checker(biny, path):
+    binaries = [
+        abspath("./checkers/%s" % (biny)),
+        abspath("/usr/lib/dpu/checkers/%s" % (biny))
+    ]
+    for binary in binaries:
+        if os.path.exists(binary):
+            null = open("/dev/null", "w")
+            stderr = null
 
-    null = open("/dev/null", "w")
-    stderr = null
-
-    cmd = [binary, path]
-    output = subprocess.check_output(cmd,
-                                     shell=False,
-                                     stderr=stderr)
+            cmd = [binary, path]
+            output = subprocess.check_output(cmd,
+                                             shell=False,
+                                             stderr=stderr)
+            return
+    raise Exception("No such builder: %s" % (biny))
 
 
 def is_identical_with_diff(from_file, to_file,
